# core/openapi.yml
openapi: 3.0.3
info:
  title: Medama Analytics API
  version: 0.1.0
  description: |
    This is the Medama Analytics OpenAPI 3.0 specification.
  contact:
    email: hello@ayuhito.com
    name: Medama Analytics
    url: https://medama.io/contact
  license:
    url: "https://github.com/ayuhito/medama/blob/main/core/LICENSE"
    name: BUSL 1.1
servers:
  - url: "http://localhost:8080"
    description: Local Server
...
paths:
  /auth/login:
    post:
      tags:
        - auth
      summary: Session token authentication.
      description: Login to the service and retrieve a session token for authentication.
      operationId: post-auth-login
      requestBody:
        description: Login details.
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
              required:
                - email
                - password
      responses:
        "200":
          description: Success
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Set the cookie for the session.
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorisedError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      servers:
        - url: "http://localhost:8080"
          description: Local Server
  /event/hit:
    post:
      tags:
        - event
      summary: Send a hit event.
      description: Send a hit event to register a user view.
      operationId: post-event-hit
      parameters:
        - name: User-Agent
          in: header
          schema:
            type: string
        - name: Accept-Language
          in: header
          description: Used to infer user language.
          schema:
            type: string
        - name: Sec-Ch-Ua
          in: header
          schema:
            type: string
        - name: Sec-Ch-Ua-Mobile
          in: header
          schema:
            type: string
        - name: Sec-Ch-Ua-Platform
          in: header
          schema:
            type: string
      requestBody:
        description: Hit event metadata.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventHit"
      responses:
        "200":
          description: OK.
        "404":
          description: Not Found
        "500":
          $ref: "#/components/responses/InternalServerError"
      servers:
        - url: "http://localhost:8080"
          description: Local Server
...
  /users:
    post:
      tags:
        - user
      summary: Create new user.
      description: Create a new user.
      operationId: post-user
      requestBody:
        description: Post the necessary fields for the API to create a new user.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: User Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserGet"
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Set the cookie for the session.
          links:
            GetUserByUserID:
              operationId: get-users-userId
              parameters:
                userId: "$response.body#/id"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorisedError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      servers:
        - url: "http://localhost:8080"
          description: Local Server
  "/users/{userId}":
    get:
      tags:
        - user
      security:
        - CookieAuth: []
      summary: Get User Info by User ID.
      description: Retrieve the information of the user with the matching user ID.
      operationId: get-users-userId
      parameters:
        - name: userId
          in: path
          description: Id of an existing user.
          required: true
          schema:
            type: string
            format: uuidv7
      responses:
        "200":
          description: User Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserGet"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      servers:
        - url: "http://localhost:8080"
          description: Local Server
    patch:
      tags:
        - user
      security:
        - CookieAuth: []
      summary: Update User Info by User ID.
      description: Update a user account's details.
      operationId: patch-users-userId
      parameters:
        - name: userId
          in: path
          description: ID of an existing user.
          required: true
          schema:
            type: string
            format: uuidv7
      requestBody:
        description: User details to update.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPatch"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserGet"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorisedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      servers:
        - url: "http://localhost:8080"
          description: Local Server
  /websites:
    get:
      tags:
        - website
      security:
        - CookieAuth: []
      summary: List of websites.
      description: Get a list of all websites from the user.
      operationId: get-websites
      parameters:
        - $ref: "#/components/parameters/SessionAuth"
      responses:
        "200":
          description: Returns a list of websites.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebsiteGet"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorisedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      servers:
        - url: "http://localhost:8080"
          description: Local Server
    post:
      tags:
        - website
      security:
        - CookieAuth: []
      summary: Add website.
      description: Add a new website.
      operationId: post-websites
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebsiteCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebsiteGet"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorisedError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      servers:
        - url: "http://localhost:8080"
          description: Local Server
  "/websites/{hostname}":
    get:
      tags:
        - website
      security:
        - CookieAuth: []
      summary: Get website details.
      description: Get website details for an individual website.
      operationId: get-websites-id
      parameters:
        - $ref: "#/components/parameters/SessionAuth"
        - $ref: "#/components/parameters/Hostname"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebsiteGet"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorisedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      servers:
        - url: "http://localhost:8080"
          description: Local Server
    patch:
      tags:
        - website
      security:
        - CookieAuth: []
      summary: Update website.
      description: Update a website's information.
      operationId: patch-websites-id
      parameters:
        - $ref: "#/components/parameters/SessionAuth"
        - $ref: "#/components/parameters/Hostname"
      requestBody:
        description: Website details to update.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebsitePatch"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebsiteGet"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorisedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      servers:
        - url: "http://localhost:8080"
          description: Local Server
    delete:
      tags:
        - website
      security:
        - CookieAuth: []
      summary: Delete website.
      description: Delete a website.
      operationId: delete-websites-id
      parameters:
        - $ref: "#/components/parameters/SessionAuth"
        - $ref: "#/components/parameters/Hostname"
      responses:
        "200":
          description: Success
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorisedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      servers:
        - url: "http://localhost:8080"
          description: Local Server
  "/websites/{hostname}/active":
    get:
      tags:
        - stats
      security:
        - CookieAuth: []
      summary: Active users.
      description: Return the number of active users who triggered a pageview in the past 5 minutes.
      operationId: get-websites-id-active
      parameters:
        - $ref: "#/components/parameters/SessionAuth"
        - $ref: "#/components/parameters/Hostname"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsActive"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorisedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      servers:
        - url: "http://localhost:8080"
          description: Local Server
  "/website/{hostname}/summary":
    get:
      tags:
        - stats
      security:
        - CookieAuth: []
      summary: Get summary of website stats.
      description: Get a summary of the website's stats.
      operationId: get-website-id-summary
      parameters:
        - $ref: "#/components/parameters/SessionAuth"
        - $ref: "#/components/parameters/PeriodStart"
        - $ref: "#/components/parameters/PeriodEnd"
        - $ref: "#/components/parameters/Hostname"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsSummary"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorisedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      servers:
        - url: "http://localhost:8080"
          description: Local Server
components:
  securitySchemes:
    CookieAuth:
      name: _me_sess
      in: cookie
      type: apiKey
      description: Session token for authentication.
  parameters:
    SessionAuth:
      name: _me_sess
      in: cookie
      description: Session token for authentication.
      required: true
      schema:
        type: string
    Hostname:
      name: hostname
      in: path
      description: Hostname for the website.
      required: true
      schema:
        type: string
    PeriodStart:
      name: start
      in: query
      description: Start time (seconds) in Unix epoch format.
      required: false
      schema:
        type: string
    PeriodEnd:
      name: end
      in: query
      description: End time (seconds) in Unix epoch format.
      required: false
      schema:
        type: string
  responses:
    NotFoundError:
      description: 404 Not Found.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                additionalProperties: false
                properties:
                  code:
                    type: integer
                    format: int32
                    default: 404
                  message:
                    type: string
                required:
                  - code
                  - message

            required:
              - error
    BadRequestError:
      description: 400 Bad Request.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                additionalProperties: false
                properties:
                  code:
                    type: integer
                    format: int32
                    default: 400
                  message:
                    type: string
                required:
                  - code
                  - message
            required:
              - error
    ConflictError:
      description: 409 Conflict Found.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                additionalProperties: false
                properties:
                  code:
                    type: integer
                    format: int32
                    default: 409
                  message:
                    type: string
                required:
                  - code
                  - message
            required:
              - error
    UnauthorisedError:
      description: 401 Unauthorised.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                additionalProperties: false
                properties:
                  code:
                    type: integer
                    format: int32
                    default: 401
                  message:
                    type: string
                required:
                  - code
                  - message
            required:
              - error
    InternalServerError:
      description: 500 Unexpected Internal Server Error.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                additionalProperties: false
                properties:
                  code:
                    type: integer
                    format: int32
                    default: 500
                  message:
                    type: string
                required:
                  - code
                  - message
            required:
              - error
  schemas:
    EventHit:
      type: object
      title: EventHit
      description: Website hit event.
      properties:
        b:
          type: string
          description: UUID generated for each user to link multiple events on the same page together.
        u:
          type: string
          description: Page URL including query parameters.
        r:
          type: string
          description: Referrer URL.
        e:
          type: string
          description: "Event type consisting of either 'pagehide', 'unload', 'load', 'hidden' or 'visible'."
        t:
          type: integer
          description: Title of page.
        d:
          type: string
          description: Timezone of the user.
        w:
          type: integer
          description: Screen width.
        h:
          type: integer
          description: Screen height.
        m:
          type: integer
          description: Time spent on page. Only sent on unload.
      required:
        - b
        - u
        - r
        - e
    UserCreate:
      type: object
      title: User
      description: User model for admin.
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        language:
          type: string
          enum: [en]
          default: en
      required:
        - email
        - password
    UserGet:
      type: object
      title: User
      description: User model for admin.
      properties:
        id:
          type: string
          description: Unique identifier for the given user.
          format: uuidv7
        email:
          type: string
          format: email
        language:
          type: string
        dateCreated:
          type: integer
          format: int64
        dateUpdated:
          type: integer
          format: int64
      required:
        - id
        - email
        - language
        - dateCreated
        - dateUpdated
    UserPatch:
      type: object
      title: User
      description: User model for admin.
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        language:
          type: string
    WebsiteGet:
      type: object
      title: Website
      description: Website details.
      properties:
        name:
          type: string
        hostname:
          type: string
          format: hostname
      required:
        - name
        - hostname
    WebsiteCreate:
      type: object
      title: Website
      description: Website details.
      properties:
        name:
          type: string
        hostname:
          type: string
      required:
        - name
        - hostname
    WebsitePatch:
      type: object
      title: Website
      description: Website details.
      properties:
        name:
          type: string
        hostname:
          type: string
    StatsActive:
      type: object
      title: StatsActive
      description: Return the number of active realtime users.
      properties:
        visitors:
          type: integer
      required:
        - visitors
    StatsSummary:
      type: object
      title: StatsSummary
      properties:
        uniques:
          type: integer
        pageviews:
          type: integer
        bounces:
          type: number
          format: float
        duration:
          type: integer